Updated some info
I updated a few bits of information on the homepage to have the names of the 2023 team for CDSpec. I also added some comments (that hopefully everyone can see) about what to do with the homepage. At the moment, the homepage seems like a glorified credits, which is going to be done by Garrett anyway.