This should get the site up.
I don't understand, and I am confused.

The issue with launching the version contained in this GitHub is that it used django version 4.0.10 and django-debug-toolbar version 3.1.1. This is a problem because django version 3.1 deprecated the WSGIRequest method `is_ajax()`, yet django-debug-toolbar version 3.1.1. still used that method.

This change only changes the django-debug-toolbar version to version 3.3.0, which no longer used `is_ajax()`.

What confuses me, is that this should, and does, make any and every application of this GitHub (or any application using django version >= 3.1 and django-debug-toolbar < 3.3.0) unusable.

So how is the website online?
https://cdspeccollab.smcm.edu/